Seperated the database from client
